code/03.Finetuning/olive-lab/requirements.txt
olive-ai==0.7.1
transformers==4.44.2
autoawq==0.2.6
optimum==1.23.1
peft==0.13.2
bitsandbytes==0.44.1
accelerate>=0.30.0
scipy==1.14.1
azure-ai-ml==1.21.1
onnxruntime-genai-cuda==0.5.0
tabulate==0.9.0
openai==1.54.4
python-dotenv==1.0.1
